+++ Makefile.pre.in
@@ -2524,6 +2524,7 @@ TESTSUBDIRS=	idlelib/idle_test \
 		test/test_multiprocessing_forkserver \
 		test/test_multiprocessing_spawn \
 		test/test_pathlib \
+                test/test_pathlib/support \
 		test/test_peg_generator \
 		test/test_pydoc \
 		test/test_pyrepl \
